backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
ollama==0.4.4
anthropic==0.43.0
aiofiles==24.1.0
ezdxf==1.3.4
Pillow==11.0.0
